feature(5): Set up automatic linter and prettier
