+++ .eslintrc.json
@@ -15,7 +15,7 @@
           "error",
           {
             "type": "attribute",
-            "prefix": "app",
+            "prefix": "oc-client",
             "style": "camelCase"
           }
         ],
@@ -23,7 +23,7 @@
           "error",
           {
             "type": "element",
-            "prefix": "app",
+            "prefix": "oc-client",
             "style": "kebab-case"
           }
         ]
