+++ angular.json
@@ -12,7 +12,7 @@
       },
       "root": "",
       "sourceRoot": "src",
-      "prefix": "app",
+      "prefix": "oc-client",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:application",
