--- src/app/Components/app/app.component.ts
@@ -1,14 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterOutlet } from '@angular/router';
-import { HeaderComponent } from '../../Layouts/Elements/Header/Components/header/header.component';
-import { MatListModule } from '@angular/material/list';
-
-@Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [CommonModule, RouterOutlet, HeaderComponent, MatListModule],
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
-})
-export class AppComponent {}
