--- src/app/Layouts/Core/MainLayout/Components/main-layout/main-layout.component.html
@@ -1,7 +1,0 @@
-<div class="min-h-screen flex flex-col">
-  <app-header></app-header>
-  <div class="flex-grow">
-    <ng-content></ng-content>
-  </div>
-  <app-footer></app-footer>
-</div>
