--- src/app/Layouts/Elements/Header/Components/header/header.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { MatIconModule } from '@angular/material/icon';
-
-@Component({
-  selector: 'app-header',
-  standalone: true,
-  imports: [MatIconModule],
-  templateUrl: './header.component.html',
-  styleUrl: './header.component.scss',
-})
-export class HeaderComponent {}
