--- src/app/Pages/NotFound/Components/not-found/not-found.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-not-found',
-  standalone: true,
-  imports: [],
-  templateUrl: './not-found.component.html',
-  styleUrl: './not-found.component.scss',
-})
-export class NotFoundComponent {}
