--- src/app/Pages/Shop/Components/shop/shop.component.html
@@ -1,3 +1,0 @@
-<app-main-layout>
-  <h1>Shop Content</h1>
-</app-main-layout>
