+++ src/app/layouts/core/MainLayout/Components/main-layout/main-layout.component.html
@@ -1,0 +1,7 @@
+<div class="min-h-screen flex flex-col">
+  <oc-client-header></oc-client-header>
+  <div class="flex-grow">
+    <ng-content></ng-content>
+  </div>
+  <oc-client-footer></oc-client-footer>
+</div>
