+++ src/app/layouts/elements/footer/components/footer/footer.component.spec.ts
