+++ src/app/layouts/elements/header/components/header/header.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { MatIconModule } from '@angular/material/icon';
+
+@Component({
+  selector: 'oc-client-header',
+  standalone: true,
+  imports: [MatIconModule],
+  templateUrl: './header.component.html',
+  styleUrl: './header.component.scss',
+})
+export class HeaderComponent {}
