+++ src/app/pages/home/components/home/home.component.html
@@ -1,0 +1,6 @@
+<oc-client-main-layout>
+  <img
+    src="https://media.wired.com/photos/650350e08c7815e31e863c7f/16:9/w_2400,h_1350,c_limit/Which-iPhone-15-Should-You-Buy-Gear.jpg"
+    alt=""
+  />
+</oc-client-main-layout>
