+++ src/app/pages/not-found/components/not-found/not-found.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'oc-client-not-found',
+  standalone: true,
+  imports: [],
+  templateUrl: './not-found.component.html',
+  styleUrl: './not-found.component.scss',
+})
+export class NotFoundComponent {}
