+++ src/app/pages/shop/components/shop/shop.component.html
@@ -1,0 +1,3 @@
+<oc-client-main-layout>
+  <h1>Shop Content</h1>
+</oc-client-main-layout>
