+++ src/app/pages/shop/components/shop/shop.component.spec.ts
