+++ src/app/pages/shop/components/shop/shop.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { MainLayoutComponent } from '../../../../Layouts/Core/MainLayout/Components/main-layout/main-layout.component';
+
+@Component({
+  selector: 'oc-client-shop',
+  standalone: true,
+  imports: [MainLayoutComponent],
+  templateUrl: './shop.component.html',
+  styleUrl: './shop.component.scss',
+})
+export class ShopComponent {}
