+++ src/index.html
@@ -16,6 +16,6 @@
     />
   </head>
   <body>
-    <app-root></app-root>
+    <oc-client-root></oc-client-root>
   </body>
 </html>
